refactored common search widget
